+++ vim/install.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
+. $(dirname $0)/../include.sh
 
-IDE="$(pwd)/ide.vim"
-CORE="$(pwd)/core.vim"
+if [ -e ~/.bashrc ]; then
+    echo "Vim is already set up"
+    exit 2
+fi
+
+echo "Setting up vim-config"
+
 VIMRC="$(pwd)/vimrc"
-
-echo "source $IDE" > $VIMRC
-echo "source $CORE" >> $VIMRC
 
 git clone https://github.com/gmarik/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 
 vim -u $VIMRC +PluginInstall +qall
 
 ln -s $VIMRC ~/.vimrc
-
